Added mistake messages
The player will see a message if he approve an article that doesn't comply with the rules. If the player makes a second mistake, he will receive a penalty.

Removed playing sounds at the beginning of the game.

Next time, the transition between the levels with the current statistics will be added.