Added main menu, losing screen, progress rollback
The player loses if:
- he didn't pay his rent for a long time;
- he didn't pay for the groceries for a long time;
- he was fired for multiple violations of the rules;

If you lose, you can return to the end of the previous level.
You can also start a new game if you go back to the main menu.

Added 2 difficulty levels: easy and normal. On the easy difficulty level, the salary is increased (if the player does everything correctly, of course).